refactor(safety): 安全类型迁移到集中式 types/security.d.ts
